Update spot-v2.js
I've updated the generateCommentElement() function to check if the logged-in user is the author of the comment and, if so, show the delete button and attach an event listener to it. When the delete button is clicked, it hides itself and shows the confirmation button. Then, when the confirmation button is clicked, it sends a DELETE request to the Xano API to delete the comment. If the deletion is successful, it removes the comment element from the DOM.

In the form submission event listener, I've also updated the new comment creation to pass in the ID returned by the Xano API (as data.id) and made the generateCommentElement function call await since it's now asynchronous. I've assumed the server's response includes the comment's id in a property named id. If it's different, please adjust accordingly.