Update spot-v2.js
The error indicates that the 'appendChild' method is trying to append something that is not a Node (element). This is happening in the displayComments function, where we are appending a comment or reply to the parent element.

This might be happening because generateCommentElement function is returning something other than a DOM Node. generateCommentElement is an async function which means it returns a Promise, not the actual Node, and you should await it.